requirements.txt
python-telegram-bot==20.7
groq==0.9.0
openai==1.12.0
aiohttp[speedups]==3.9.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
pytz==2024.1
beautifulsoup4==4.12.3
apscheduler==3.10.4
feedparser==6.0.11
pymupdf==1.24.0
